chore: P2-T4 - 字节级清理旧播放代码与测试桩
